requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
openai==0.27.0
tiktoken==0.3.0
python-dotenv==0.19.0
httpx==0.23.0
aiofiles==0.8.0
jinja2==3.0.3
starlette==0.16.0